mattdrayer/api-migrate_transitional_ids: Cleaned up two additional tables
merged with master

changes after matts suggestions
