+++ lms/djangoapps/api_manager/courses/serializers.py
@@ -1,19 +1,7 @@
 """ Django REST Framework Serializers """
 
-from api_manager.models import CourseModuleCompletion
 from api_manager.utils import generate_base_uri
 from rest_framework import serializers
-
-
-class CourseModuleCompletionSerializer(serializers.ModelSerializer):
-    """ Serializer for CourseModuleCompletion model interactions """
-    user_id = serializers.Field(source='user_id')
-
-    class Meta:
-        """ Serializer/field specification """
-        model = CourseModuleCompletion
-        fields = ('id', 'user_id', 'course_id', 'content_id', 'stage', 'created', 'modified')
-        read_only = ('id', 'created')
 
 
 class GradeSerializer(serializers.Serializer):
