+++ lms/djangoapps/api_manager/management/commands/tests/test_migrate_stage_prefix.py
@@ -9,7 +9,7 @@ from django.contrib.auth.models import Group, User
 from django.test import TestCase
 from django.test.utils import override_settings
 
-from api_manager import models as api_models
+from progress.models import CourseModuleCompletion
 from api_manager.management.commands import migrate_stage_prefix
 from courseware.tests.modulestore_config import TEST_DATA_MIXED_MODULESTORE
 from xmodule.modulestore.tests.factories import CourseFactory, ItemFactory
@@ -40,19 +40,19 @@ class MigrateCourseIdsTests(TestCase):
         """
         # Set up the data to be migrated
         user = User.objects.create(email='testuser@edx.org', username='testuser', password='testpassword', is_active=True)
-        course_module_completion = api_models.CourseModuleCompletion.objects.create(user=user, course_id=self.good_style_course_id, content_id=self.good_style_content_id, stage=self.bad_style_stage)
+        course_module_completion = CourseModuleCompletion.objects.create(user=user, course_id=self.good_style_course_id, content_id=self.good_style_content_id, stage=self.bad_style_stage)
 
         user2 = User.objects.create(email='testuser2@edx.org', username='testuser2', password='testpassword2', is_active=True)
-        course_module_completion2 = api_models.CourseModuleCompletion.objects.create(user=user2, course_id=self.good_style_course_id2, content_id=self.good_style_content_id2, stage=self.bad_style_stage2)
+        course_module_completion2 = CourseModuleCompletion.objects.create(user=user2, course_id=self.good_style_course_id2, content_id=self.good_style_content_id2, stage=self.bad_style_stage2)
 
 
         # Run the data migration
         migrate_stage_prefix.Command().handle()
 
 
-        updated_course_module_completion = api_models.CourseModuleCompletion.objects.get(id=course_module_completion.id)
+        updated_course_module_completion = CourseModuleCompletion.objects.get(id=course_module_completion.id)
         self.assertEqual(updated_course_module_completion.stage, self.good_style_stage)
 
-        updated_course_module_completion = api_models.CourseModuleCompletion.objects.get(id=course_module_completion2.id)
+        updated_course_module_completion = CourseModuleCompletion.objects.get(id=course_module_completion2.id)
         self.assertEqual(updated_course_module_completion.stage, self.good_style_stage2)
         print "Course Module Completion Data Migration Passed"
