+++ lms/djangoapps/api_manager/models.py
@@ -133,28 +133,6 @@ class CourseContentGroupRelationship(TimeStampedModel):
         unique_together = ("course_id", "content_id", "group_profile")
 
 
-class CourseModuleCompletion(TimeStampedModel):
-    """
-    The CourseModuleCompletion model contains user, course, module information
-    to monitor a user's progression throughout the duration of a course,
-    we need to observe and record completions of the individual course modules.
-    """
-    user = models.ForeignKey(User, db_index=True, related_name="course_completions")
-    course_id = models.CharField(max_length=255, db_index=True)
-    content_id = models.CharField(max_length=255, db_index=True)
-    stage = models.CharField(max_length=255, null=True, blank=True)
-
-    @classmethod
-    def get_actual_completions(cls):
-        """
-        This would skip those modules with ignorable categories
-        """
-        detached_categories = getattr(settings, 'PROGRESS_DETACHED_CATEGORIES', [])
-        cat_list = [Q(content_id__contains=item.strip()) for item in detached_categories]
-        cat_list = reduce(lambda a, b: a | b, cat_list)
-        return cls.objects.all().exclude(cat_list)
-
-
 class APIUserQuerySet(models.query.QuerySet):  # pylint: disable=R0924
     """ Custom QuerySet to modify id based lookup """
     def filter(self, *args, **kwargs):
